fix gsap issues
