+++ app/layout.js
@@ -5,73 +5,12 @@ import theme from "@config/theme.json";
 import TwSizeIndicator from "@layouts/components/TwSizeIndicator";
 import Footer from "@layouts/partials/Footer";
 import Header from "@layouts/partials/Header";
-import { gsap } from "@lib/gsap";
-import { useEffect, useRef } from "react";
 import "../styles/style.scss";
 
 export default function RootLayout({ children }) {
   // import google font css
   const pf = theme.fonts.font_family.primary;
   const sf = theme.fonts.font_family.secondary;
-  const main = useRef();
-
-  //gsap fade animation
-  useEffect(() => {
-    const ctx = gsap.context(() => {
-      //fade
-      const fadeElements = document.querySelectorAll(".fade");
-      fadeElements.forEach((el) => {
-        gsap.to(el, {
-          opacity: 1,
-          scrollTrigger: el,
-          duration: 0.3,
-        });
-      });
-
-      //gsap animation
-      const elements = document.querySelectorAll(".animate");
-      elements.forEach((el) => {
-        const tl = gsap.timeline({
-          scrollTrigger: {
-            trigger: el,
-            start: "top bottom",
-            // markers: true,
-          },
-        });
-
-        if (el.classList.contains("from-left")) {
-          tl.from(el, {
-            opacity: 0,
-            x: -100,
-          });
-        } else if (el.classList.contains("from-right")) {
-          tl.from(el, {
-            opacity: 0,
-            x: 100,
-          });
-        } else {
-          tl.from(el, {
-            opacity: 0,
-            y: 100,
-          });
-        }
-      });
-
-      //background animation
-      const animatedBgs = document.querySelectorAll(".bg-theme");
-      animatedBgs.forEach((bg) => {
-        gsap.to(bg, {
-          scrollTrigger: {
-            trigger: bg,
-            toggleClass: "bg-animate",
-            once: true,
-          },
-        });
-      });
-    }, main);
-
-    return () => ctx.revert();
-  }, []);
 
   return (
     <html suppressHydrationWarning={true} lang="en">
@@ -117,7 +56,7 @@ export default function RootLayout({ children }) {
       <body suppressHydrationWarning={true}>
         <TwSizeIndicator />
         <Header />
-        <main ref={main}>{children}</main>
+        {children}
         <Footer />
       </body>
     </html>
