+++ layouts/Default.js
@@ -2,7 +2,6 @@
 
 import shortcodes from "@shortcodes/all";
 import dynamic from "next/dynamic";
-import { useEffect, useState } from "react";
 import Banner from "./components/Banner";
 
 const MDXRemote = dynamic(
@@ -15,18 +14,13 @@ const MDXRemote = dynamic(
 const Default = ({ data }) => {
   const { frontmatter, mdxContent } = data;
   const { title } = frontmatter;
-  const [isMounted, setIsMounted] = useState(false);
-
-  useEffect(() => {
-    setIsMounted(true);
-  }, []);
 
   return (
     <section className="section">
       <Banner title={title} />
       <div className="container mt-10">
         <div className="content">
-          {isMounted && <MDXRemote {...mdxContent} components={shortcodes} />}
+          <MDXRemote {...mdxContent} components={shortcodes} />
         </div>
       </div>
     </section>
